+++ src/cli/commands/anon-col/anon-col/command.ts
@@ -1,6 +1,6 @@
 import {Command, Option} from 'commander';
-import {providers} from '../../../types/types';
-import {anonColAction} from './anon-col-action';
+import {providers} from '../../../../types/types';
+import {maskCommand} from '../commands/mask/mask-command';
 
 export const anonColCommand = new Command('anon-col');
 
@@ -10,10 +10,11 @@ anonColCommand
 	.argument('<db>', 'database name')
 	.argument('<table>', 'table name')
 	.argument('<column>', 'column name')
+	.addCommand(maskCommand, {isDefault: true})
 	.addOption(
 		new Option('-p --provider <provider>', 'provider to be used for column')
 			.default('mask')
 			.choices(providers),
 	)
-	.option('-Confirm --no-confirm', 'Do not wait for user confirmation')
-	.action(anonColAction);
+	.option('-skip --no-confirm', 'Do not wait for user confirmation');
+// .action(anonColAction);
