+++ src/cli/dbzar.ts
@@ -2,7 +2,7 @@
 
 import process from 'node:process';
 import {Command} from 'commander';
-import {anonColCommand} from './commands/anon-col/anon-col-command';
+import {anonColCommand} from './commands/anon-col/anon-col/command';
 
 const program = new Command();
